Breaking out AbstractSettlingCamera between AbstractCamera and ReferenceCamera
- Moving ReferenceCamera from reference to reference.camera
- Moving AbstractBroadcastingCamera from reference to reference.camera
- Fixing a bug, where large Gaussian blur image rescaling did not
actually work.
