+++ src/main/java/org/openpnp/gui/wizards/CameraConfigurationWizard.java
@@ -51,10 +51,10 @@ import org.openpnp.gui.support.LengthConverter;
 import org.openpnp.gui.support.MessageBoxes;
 import org.openpnp.gui.support.MutableLocationProxy;
 import org.openpnp.gui.support.NamedConverter;
-import org.openpnp.machine.reference.AbstractBroadcastingCamera;
-import org.openpnp.machine.reference.ReferenceCamera;
-import org.openpnp.machine.reference.ReferenceCamera.FocusSensingMethod;
 import org.openpnp.machine.reference.axis.ReferenceVirtualAxis;
+import org.openpnp.machine.reference.camera.AbstractBroadcastingCamera;
+import org.openpnp.machine.reference.camera.ReferenceCamera;
+import org.openpnp.machine.reference.camera.ReferenceCamera.FocusSensingMethod;
 import org.openpnp.model.Configuration;
 import org.openpnp.model.Length;
 import org.openpnp.model.LengthUnit;
