+++ src/main/java/org/openpnp/machine/neoden4/Neoden4SwitcherCamera.java
@@ -6,7 +6,7 @@ import java.util.Map;
 
 import org.openpnp.gui.support.Wizard;
 import org.openpnp.machine.neoden4.wizards.Neoden4SwitcherCameraConfigurationWizard;
-import org.openpnp.machine.reference.ReferenceCamera;
+import org.openpnp.machine.reference.camera.ReferenceCamera;
 import org.openpnp.model.Configuration;
 import org.openpnp.spi.Camera;
 import org.openpnp.spi.PropertySheetHolder;
