+++ src/main/java/org/openpnp/machine/reference/camera/calibration/AdvancedCalibration.java
@@ -33,8 +33,8 @@ import org.opencv.core.Point;
 import org.opencv.core.Point3;
 import org.opencv.core.Scalar;
 import org.opencv.core.Size;
-import org.openpnp.machine.reference.ReferenceCamera;
 import org.openpnp.machine.reference.ReferenceHead;
+import org.openpnp.machine.reference.camera.ReferenceCamera;
 import org.openpnp.model.Length;
 import org.openpnp.model.LengthUnit;
 import org.openpnp.model.Location;
